Cria controllers indiviualizados para 'listarVendas' e 'cadastrarVenda'
