+++ routes/api.php
@@ -3,7 +3,8 @@
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\VendedorController;
-use App\Http\Controllers\VendasController;
+use App\Http\Controllers\CadastrarVendaController;
+use App\Http\Controllers\ListarVendasController;
 
 /*
 |--------------------------------------------------------------------------
@@ -24,11 +25,11 @@ Route::post('/cadastrarVendedor', [VendedorController::class, 'cadastrarVendedor
 
 Route::get('/listarVendedores', [VendedorController::class, 'listarVendedores']);
 
-Route::post('/cadastrarVenda', [VendasController::class, 'cadastrarVenda']);
+Route::post('/cadastrarVenda', [CadastrarVendaController::class, 'cadastrarVenda']);
 
-Route::get('/listarVendas', [VendasController::class, 'listarVendas']);
+Route::get('/listarVendas', [ListarVendasController::class, 'listarVendas']);
 
-Route::get('/listarVendas/Vendedor/{vendedor_id}', [VendasController::class, 'listarVendasVendedor']);
+Route::get('/listarVendas/Vendedor/{vendedor_id}', [ListarVendasController::class, 'listarVendasVendedor']);
 
 Route::delete('/deletarVendedor/{vendedor_id}', [VendedorController::class, 'deletarVendedor']);
